requirements.txt
flask==3.0.3
flask-cors==4.0.0
Pillow==10.3.0
numpy==1.26.4
huggingface_hub==0.23.0
groq==0.8.0
